+++ EssentialFeed/Feed Cache/FeedStore.swift
@@ -1,0 +1,16 @@
+//
+//  FeedStore.swift
+//  EssentialFeed
+//
+//  Created by Ana Anguiano Cruz on 18/11/21.
+//
+
+import Foundation
+
+public protocol FeedStore {
+    typealias DeletionCompletion = (Error?) -> Void
+    typealias InsertionCompletion = (Error?) -> Void
+    
+    func deleteCachedFeed(completion: @escaping DeletionCompletion)
+    func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
+}
